fix(channels): address PR #245 re-review findings (Med/Low)
Applies four findings from the second-pass review of PR #245.  Each
change is TDD: the new/updated tests were committed to red first, then
the production fix was written to make them green.

ReconcileConfig — atomic create (Med, Must-Fix #2)
  internal/channels/router.go: the missing-channel arm of
  ReconcileConfig was still calling CreateChannel + N×AddMember outside
  any transaction.  A failure mid-loop (transient store error or an
  invalid declared member that bypassed Config.Validate) left the
  channel row committed with only a prefix of the declared membership;
  the next startup then tripped ErrConfigStoreMembershipDivergence and
  required manual operator cleanup.  Mirrors the handler-side atomicity
  fix from the v1 review: now calls CreateChannelWithMembers in one
  shot.  Drops the dead `_ = existing` discard on the same path (Nit).
  Pinned by: TestChannelRouter_ReconcileConfig_AtomicOnInvalidMember.

Invalid channel name → 400 not 500 (Low/Med)
  internal/channels/sqlite.go: the store's name-required and
  name-pattern errors in both CreateChannel and CreateChannelWithMembers
  were plain fmt.Errorf / errors.New.  writeChannelError's default case
  maps unknown errors to 500 INTERNAL, so an operator typing
  `Name: "Sprint 1"` got a 500 instead of the correct 400.  Wrapped
  both with ErrInvalidChannelType (the closest existing 400-class
  sentinel — these ARE invalid channel attributes) so the HTTP status
  is now correct without touching the human-readable message text.
  Pinned by: TestChannels_CreateChannel_InvalidName_400 (table-driven:
  uppercase / whitespace / path-traversal shapes).

Trust-boundary callout (Must-Fix #1)
  cmd/orchestrator/channels.go: emits a one-shot Warn at startup
  whenever the channels subsystem is enabled, surfacing the v0.3.0
  unauthenticated REST surface in the operator's first log scrape.
  Intentionally not opt-out-able (an opt-out would defeat the warning).
  CHANGELOG.md: new RFC 0011 PR 2 entry under [Unreleased] with an
  explicit "Security — UNAUTHENTICATED in v0.3.0" paragraph that
  documents the trust boundary and the recommended mitigations
  (127.0.0.1 bind / reverse proxy / firewall) for operators.

Skipped (with justification):
  - Body-size cap: already enforced via decodeJSON's 1 MiB
    MaxBytesReader wrapper — review missed this.
  - respond boundary validation: externally observable behaviour is
    already 400 via ErrInvalidRespondPolicy sentinel path in
    writeChannelError; existing tests pin it; handler duplication would
    only save one DB round-trip.
  - ISSUE-0008 created date: already 2026-05-04 in the file.
  - Router-nil fallback Warn / list cursor / divergence Warn / history
    probe round-trip: deferred per review's own "Nice to Have" and
    "OQ-7 deferred to PR 7" classifications.
