+++ FILEMAP.md
@@ -1,13 +1,13 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-05-04  
-> **618** tracked files
+> **619** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
-| `internal/` | 180 |
+| `internal/` | 181 |
 | `tests/` | 114 |
 | `docs/` | 112 |
 | `agents/` | 93 |
@@ -28,7 +28,7 @@
 | Extension | Count |
 |-----------|-------|
 | `.py` | 217 |
-| `.go` | 183 |
+| `.go` | 184 |
 | `.md` | 149 |
 | `.yaml` | 27 |
 | `.rs` | 9 |
@@ -369,6 +369,7 @@
 │   │   ├── sqlite_messages.go
 │   │   ├── sqlite_pr2_review_test.go
 │   │   ├── sqlite_pr2_test.go
+│   │   ├── sqlite_query.go
 │   │   ├── sqlite_schema.go
 │   │   ├── sqlite_test.go
 │   │   ├── store.go
